Comments upto 002, adding 003.

+++ rust/projecteuler_001.rs
@@ -18,15 +18,22 @@ https://www.rust-lang.org/
 
 fn main() {
 
-	let mut c = 3i64;
+	//Define the starting number 32-bit integer
+	//Starting number is 3 since 1 or 2 do not matter (multiples start at 3)
+	let mut c = 3i32;
+
+	//define the total sum as 64-bit integer to fit a larger number
 	let mut total = 0i64;
 
+	//While c is less than 1000 (incrememting by 1 each iteration)
+	//check if c is divisible by 3 or 5, and add to the total
 	while c < 1000 {
 		if c % 3 == 0 || c % 5 == 0 {
-			total += c;
+			total += c as i64;
 		}
 		c += 1;
 	}
 
+	//Print statement for the answer. {} is a placeholder in rust
 	println!("Sum of multiples of 3 and 5 less than 1000:\n{}\n", total);
 }
